requirements.txt
agno>=1.0.0
rich>=13.7.0
PyYAML>=6.0.1
typer>=0.12.5
colorama>=0.4.6
# worflows dependencies
ollama
yfinance
ddgs